LayoutTests/media/encrypted-media/clearKey/clearKey-cenc-stop-playback-before-key.html

<!DOCTYPE html>
<html>
<head>
    <title>Playback ClearKey CENC audio with full-sample encryption</title>
    <script src="../medias-enc.js"></script>
    <script src="../../video-test.js"></script>
    <script src="../../media-source/media-source-loader-simple.js"></script>
    <script src="encrypted-media-clearKey-handler.js"></script>
    <script>

     const audioConf = streamMedias["simpleClearKeyMSE"].audio;

     function runTest() {

         findMediaElement();

         const emeHandler = new EncryptedMediaHandler(video, audioConf);
         if (!emeHandler)
             failTest();

         emeHandler.skipUpdateCallback = function() {
             setTimeout(function() {
                 logResult(true, "Changing URL");
                 const startTime = Date.now();
                 video.src = "../../content/counting.mp4"
                 video.play(1);
                 const difference = Date.now() - startTime;
                 logResult(difference < 4000, "URL changed in time");
                 endTest();
             }, 2000);
         }

         const ms = new MediaSourceLoaderSimple(video);
         ms.onready = function() {
             logResult(true, "Media source is opened");
             ms.createSourceBuffer(audioConf, 4);
             video.play().catch(function(e) { });
         };
     }
    </script>
</head>
<body onload="runTest()">
    <video></video>
</body>
</html>
